chore: begin ts conversion
