+++ js/package.json
@@ -7,8 +7,8 @@
     "flarum-webpack-config": "^2.0.0",
     "twemoji": "^14.0.2",
     "twemoji-basename": "^1.0.0",
-    "webpack": "^5.76.0",
-    "webpack-cli": "^4.9.2",
+    "webpack": "^5.76.2",
+    "webpack-cli": "^5.0.1",
     "webpack-merge": "^5.8.0"
   },
   "scripts": {
@@ -21,6 +21,6 @@
   "devDependencies": {
     "@flarum/prettier-config": "^1.0.0",
     "flarum-tsconfig": "^1.0.2",
-    "prettier": "^2.6.2"
+    "prettier": "^2.8.6"
   }
 }
