+++ js/src/common/utils/index.ts
