+++ js/src/forum/components/index.ts
